indexBuffer修改为使用uint16
(cherry picked from commit 265b7ad83968991c880fdf3aa48368ea2008a0a5)

+++ Source/LGUI/Public/Core/LGUIMesh/LGUIMeshComponent.h
@@ -8,7 +8,7 @@
 struct FLGUIMeshSection
 {
 	TArray<FVector> vertices;
-	TArray<uint32> triangles;
+	TArray<uint16> triangles;
 	TArray<FVector2D> uvs;
 	TArray<FVector2D> uvs1;
 	TArray<FVector2D> uvs2;
